[BC Break] changes return type of Sequence::get() to plain value
The reasoning behind this change is that if you operate on sequences,
you never pull any arbitrary index from a sequence, but either use
the implicit iterator in a FOREACH loop, or iterate over a range in
a FOR loop. As such the additional Option::get calls are unnecessary
for sequences in connection with the Sequence::get operation, and do
not add any value.

    foreach ($seq as $elem) { }
    for ($i=0,$c=count($seq); $i<$c; $i++) { $elem = $seq->get($i); }

The behavior for Maps remains unchanged as the key has a specific
meaning there, and pulling a specific key from a Map is a common
operation.